Extract login business logic
